Improve sass implementation

+++ tests/Sassin.MSTest/Tests/CompilationTest.cs
@@ -1,3 +1,4 @@
+using Acklann.Diffa;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Shouldly;
 using System;
@@ -5,6 +6,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Text;
+using System.Text.RegularExpressions;
 
 namespace Acklann.Sassin.Tests
 {
@@ -12,12 +14,13 @@ namespace Acklann.Sassin.Tests
     public class CompilationTest
     {
         [ClassInitialize]
-        public static void Initialize(TestContext context)
+        public static void Initialize(TestContext _)
         {
-            foreach (var item in Directory.EnumerateFiles(NodeJS.InstallationDirectory, "*.js"))
-            {
-                File.Delete(item);
-            }
+            if (Directory.Exists(NodeJS.InstallationDirectory))
+                foreach (var item in Directory.EnumerateFiles(NodeJS.InstallationDirectory, "*.js"))
+                {
+                    File.Delete(item);
+                }
 
             NodeJS.Install();
         }
@@ -53,7 +56,37 @@ namespace Acklann.Sassin.Tests
             totalFiles.ShouldBe(expectedFiles);
             result.GeneratedFiles.Length.ShouldBe(expectedFiles);
 
-            //Diff.Approve(builder);
+            Diff.Approve(builder, ".txt", label);
+        }
+
+        [DataTestMethod]
+        [DynamicData(nameof(GetInvalidFiles), DynamicDataSourceType.Method)]
+        public void Can_detect_sass_errors(string documentPath, int errorLine)
+        {
+            // Arrange
+            var cwd = Path.Combine(AppContext.BaseDirectory, "generated", "errors");
+            if (Directory.Exists(cwd)) Directory.Delete(cwd, recursive: true);
+            Directory.CreateDirectory(cwd);
+
+            var options = new CompilerOptions
+            {
+                Minify = false,
+                OutputDirectory = cwd,
+                GenerateSourceMaps = false,
+            };
+
+            // Act
+            var result = Sass.Compile(documentPath, options);
+            var error = result.Errors.FirstOrDefault();
+
+            // Assert
+            result.Success.ShouldBeFalse();
+            result.Errors.ShouldNotBeEmpty();
+
+            error.Line.ShouldBe(errorLine);
+            error.Column.ShouldBeGreaterThan(0);
+            error.StatusCode.ShouldBeGreaterThan(0);
+            error.Message.ShouldNotBeNullOrEmpty();
         }
 
         // ==================== DATA ==================== //
@@ -62,11 +95,38 @@ namespace Acklann.Sassin.Tests
         {
             yield return new object[]{"css", 1, new CompilerOptions
             {
+                Minify = false,
+                AddSourceComments = false,
+                GenerateSourceMaps = false
+            }};
+
+            yield return new object[]{"css-map", 2, new CompilerOptions
+            {
+                Minify = false,
+                Suffix = string.Empty,
+                AddSourceComments = true,
+                GenerateSourceMaps = true
+            }};
+
+            yield return new object[]{"css-min-map", 2, new CompilerOptions
+            {
                 Minify = true,
                 AddSourceComments = false,
-                GenerateSourceMaps = true,
-                KeepIntermediateFiles = false
+                GenerateSourceMaps = true
             }};
+        }
+
+        private static IEnumerable<object[]> GetInvalidFiles()
+        {
+            var pattern = new Regex(@"\d+");
+            foreach (string file in Directory.EnumerateFiles(Sample.DirectoryName, "error-*.scss"))
+            {
+                Match match = pattern.Match(Path.GetFileNameWithoutExtension(file));
+                if (match.Success && int.TryParse(match.Value, out int lineNo))
+                {
+                    yield return new object[] { file, lineNo };
+                }
+            }
         }
     }
 }
